CHORE: Made little changes to the entity
