+++ src/main/java/com/iamstevol/vehicleparkingsystem/entity/User.java
@@ -1,12 +1,28 @@
 package com.iamstevol.vehicleparkingsystem.entity;
 
+import com.iamstevol.vehicleparkingsystem.security.auth.ApplicationUserRole;
+import com.iamstevol.vehicleparkingsystem.security.token.Token;
 import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.Hibernate;
 import org.hibernate.annotations.GenericGenerator;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 
-import java.util.HashSet;
-import java.util.Set;
+import java.util.Collection;
+import java.util.List;
+import java.util.Objects;
 
-public class User {
+@Getter
+@Setter
+@ToString
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "_user")
+public class User implements UserDetails {
 
     @Id
     @GeneratedValue(generator = "UUID")
@@ -22,13 +38,63 @@ public class User {
 
     @Column(name = "lastname", length = 100)
     private String lastName;
+    @Column(name = "email", length = 100)
+    private String email;
+    @Column(name = "password", length = 100)
+    private String password;
 
-    @ManyToMany(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
-    @JoinTable(name = "user_roles",
-            joinColumns = {
-                    @JoinColumn(name = "user_id", referencedColumnName = "user_id")
-            },
-            inverseJoinColumns = {
-                    @JoinColumn(name = "role_id", referencedColumnName = "role_id") })
-    private Set<Role> roles = new HashSet<>();
+    @Enumerated(EnumType.STRING)
+    private ApplicationUserRole role;
+
+    @OneToMany(mappedBy = "user")
+    @ToString.Exclude
+    private List<Token> tokens;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority(role.name()));
+    }
+
+    @Override
+    public String getPassword() {
+        return password;
+    }
+
+    @Override
+    public String getUsername() {
+        return email;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
+        User user = (User) o;
+        return userId != null && Objects.equals(userId, user.userId);
+    }
+
+    @Override
+    public int hashCode() {
+        return getClass().hashCode();
+    }
 }
