+++ src/main/java/com/iamstevol/vehicleparkingsystem/repository/UserRepository.java
@@ -1,11 +1,11 @@
 package com.iamstevol.vehicleparkingsystem.repository;
 
-import com.iamstevol.vehicleparkingsystem.entity.user.User;
+import com.iamstevol.vehicleparkingsystem.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, String> {
 
   Optional<User> findByEmail(String email);
