+++ src/main/java/com/iamstevol/vehicleparkingsystem/security/token/TokenRepository.java
@@ -10,10 +10,10 @@ public interface TokenRepository extends JpaRepository<Token, Integer> {
 
   @Query(value = """
       select t from Token t inner join User u\s
-      on t.user.id = u.id\s
-      where u.id = :id and (t.expired = false or t.revoked = false)\s
+      on t.user.userId = u.userId\s
+      where u.userId = :id and (t.expired = false or t.revoked = false)\s
       """)
-  List<Token> findAllValidTokenByUser(Integer id);
+  List<Token> findAllValidTokenByUser(String id);
 
   Optional<Token> findByToken(String token);
 }
